+++ app/src/main/res/layout/alarm_listview.xml
@@ -13,8 +13,7 @@
             android:id="@+id/memberImage"
             android:layout_width="40dp"
             android:layout_height="40dp"
-            android:layout_margin="10dp"
-            android:src="@drawable/login_naver"/>
+            android:layout_margin="10dp" />
 
         <LinearLayout
             android:layout_width="match_parent"
